ADD: functions

+++ src/libs/SbtLib.sol
@@ -13,7 +13,7 @@ library SbtLib {
         mapping(address => uint8) favo;
         mapping(address => uint16) received_favo;
         mapping(address => uint16) maki;
-        mapping(address => uint16) grade;
+        mapping(address => uint8) grade;
         mapping(address => uint8) referral;
         mapping(bytes4 => bool) interfaces;
     }
